factoring

+++ app/src/main/java/com/test/myhealth/data/storage/modelDb/ArticleEntity.kt
@@ -10,6 +10,6 @@ data class ArticleEntity(
     val title: String?,
     val content: String?,
     val img: String?,
-    val favorites: Int?,
-    val likes: Int?,
+    val favorites: Long?,
+    val likes: Long?,
 )
